refactor(server): extract API and channel routes from server.ts
Split server.ts (522→~210 lines) into:
- api-routes.ts: middleware, /api/open, /api/upload
- channel-routes.ts: /api/channel-*, /api/launch-claude, /api/events

No behavior changes. Factory pattern passes Express app + middleware.
